requirements.txt
Babel==1.3
Flask==0.10.1
Flask-Babel==0.8
Flask-Login==0.2.10
Flask-Mail==0.9.0
Flask-OpenID==1.2.1
Flask-Principal==0.4.0
Flask-SQLAlchemy==0.16
Flask-Security==1.7.1
Flask-WTF==0.9.5
Flask-WhooshAlchemy==0.55a
Jinja2==2.7.2
MarkupSafe==0.19
SQLAlchemy==0.7.9
Tempita==0.5.2
WTForms==1.0.5
Werkzeug==0.9.4
Whoosh==2.6.0
blinker==1.3
decorator==3.4.0
flup==1.0.2
itsdangerous==0.24
passlib==1.6.2
python-openid==2.2.5
pytz==2013b
speaklater==1.3
sqlalchemy-migrate==0.7.2
wsgiref==0.1.2
